# Improve performance by using MULTI and MGET commands