- Using MULTI transaction for INCR + EXPIRE will guarantee that we won't ever create any keys without expiry
- Using MGET to fetch both previous and current window counters in a single atomic operation will improve throughput