@VojtechVitek 2024-07-25
-22.97% .. 1.3x speedup in the CI  🚀 

-52.53% .. 2x speedup on my macOS M2 laptop 🚀 🚀 
```
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/go-chi/httprate-redis
               │   old.txt    │               new.txt                │
               │    sec/op    │    sec/op     vs base                │
LocalCounter-8   1699.3m ± 7%   806.7m ± 11%  -52.53% (p=0.000 n=10)

               │   old.txt    │               new.txt               │
               │     B/op     │     B/op      vs base               │
LocalCounter-8   32.84Mi ± 0%   34.92Mi ± 0%  +6.32% (p=0.000 n=10)

               │   old.txt   │              new.txt               │
               │  allocs/op  │  allocs/op   vs base               │
LocalCounter-8   1.261M ± 0%   1.230M ± 0%  -2.45% (p=0.000 n=10)
```